Match-id-43fae700a959a0519bd6705c2b89b7309a8e7d51

+++ ait_backend/utils/file.cpp
@@ -152,7 +152,8 @@ bool IsPathCharactersValid(const std::string& path)
 
 bool IsPathDepthValid(const std::string& path)
 {
-    return std::count(path.begin(), path.end(), MsConst::PATH_SEPARATOR) <= MsConst::PATH_DEPTH_MAX;
+    return std::count(path.begin(), path.end(), MsConst::PATH_SEPARATOR) <=
+           static_cast<size_t>(MsConst::PATH_DEPTH_MAX);
 }
 
 bool IsRegularFile(const std::string& path)
